ai-services/requirements.txt
flask==3.0.0
flask-cors==4.0.0
openai-whisper==20231117
numpy==1.24.3
torch==2.1.0
torchaudio==2.1.0
soundfile==0.12.1
werkzeug==3.0.1
gunicorn==21.2.0